fix my previous blocking fix
daemon is failing to start because it is attempting to connect to the
containerd socker before it is available.

this fix-for-the-fix refactors endpointconn.Get to accept varying array of grpc.DialOption
which allows the daemon to block/retry while dialing containerd socket but client
connections will not block if they are lacking perms on the k3c socket.